Pull findRoot out into libZipOps
The findRoot function useful for managing mod zip files on the master.
