+++ test/lib/factorio/patch.js
@@ -8,47 +8,6 @@ const libFactorio = require("@clusterio/lib/factorio");
 
 
 describe("lib/factorio", function() {
-	describe("findRoot()", function() {
-		it("should find the root of a zip", function() {
-			let zip = new JSZip();
-			zip.file("root/foo.txt", "");
-			zip.file("root/bar.txt", "");
-
-			assert.equal(libFactorio._findRoot(zip), "root");
-		});
-
-		it("should throw if files are at the root of the zip", function() {
-			let zip = new JSZip();
-			zip.file("root/foo.txt", "");
-			zip.file("bar.txt", "");
-
-			assert.throws(
-				() => libFactorio._findRoot(zip),
-				new Error("Zip contains file 'bar.txt' in root dir")
-			);
-		});
-
-		it("should throw if there are multiple root dirs", function() {
-			let zip = new JSZip();
-			zip.file("root-1/foo.txt", "");
-			zip.file("root-2/bar.txt", "");
-
-			assert.throws(
-				() => libFactorio._findRoot(zip),
-				new Error("Zip contains multiple root folders")
-			);
-		});
-
-		it("should throw if given an empty zip file", function() {
-			let zip = new JSZip();
-
-			assert.throws(
-				() => libFactorio._findRoot(zip),
-				new Error("Empty zip file")
-			);
-		});
-	});
-
 	describe("generateLoader()", function() {
 		let reference = [
 			"-- Auto generated scenario module loader created by Clusterio",
